requirements.txt
google-genai
pyyaml
requests
pydantic
streamlit